+++ SwiftyRSA/Key.swift
@@ -81,7 +81,7 @@ public protocol Key {
         self.tag = tag
         
         self.originalData = data
-        let dataWithoutHeader = try SwiftyRSA.stripPublicKeyHeader(keyData: data)
+        let dataWithoutHeader = try SwiftyRSA.stripKeyHeader(keyData: data)
         
     	reference = try SwiftyRSA.addKey(dataWithoutHeader, isPublic: true, tag: tag)
     }
@@ -249,7 +249,8 @@ public protocol Key {
         self.originalData = data
         let tag = UUID().uuidString
         self.tag = tag
-        reference = try SwiftyRSA.addKey(data, isPublic: false, tag: tag)
+        let dataWithoutHeader = try SwiftyRSA.stripKeyHeader(keyData: data)
+        reference = try SwiftyRSA.addKey(dataWithoutHeader, isPublic: false, tag: tag)
     }
     
     /// Creates a private key with a base64-encoded string.